refactor(storage): extract table creation blocks from createSchema()
Split the monolithic 215-line SQL string in createSchema() into
domain-grouped constants in a new schema_tables.go file:
- schemaCore: tenants, session_messages, tenant_state
- schemaMemory: long_term_memory, core_memory_blocks, archival_memory
- schemaEvents: event_history + FTS trigger
- schemaUsers: user_profiles, web_users, user_settings, user_llm_configs, user_chats
- schemaRunners: runner_tokens, runners
- schemaShared: shared_registry
- schemaJobs: cron_jobs, event_triggers

createSchema() is now a concise 18-line coordinator concatenating these constants.
